Extracted common stuff in activities to a abstract base activity
